POMDPDiscrete

Documentation for POMDPDiscrete.

POMDPDiscrete.action_distribution

POMDPDiscrete.build_probabilistic_model

POMDPDiscrete.initialise_vector

POMDPDiscrete.policy_evaluation

POMDPDiscrete.policy_grid

POMDPDiscrete.policy_transition_matrix

POMDPDiscrete.random_stochastic_policy

POMDPDiscrete.uniform_stochastic_policy

POMDPDiscrete.value_iteration

POMDPs.reward

POMDPs.transition

POMDPDiscrete.action_distribution — Method

action_distribution(policy::StochasticPolicy, x)

Returns the distribution of the ordered actions given the policy and the current state or belief x.

POMDPDiscrete.build_probabilistic_model — Method

build_probabilistic_model(mdp)

Return a 3 dimensional matrix representing P[s′, a, s], taking into account successor state distributions for action/state pairs.

POMDPDiscrete.initialise_vector — Method

initialise_vector(size; rng)

Returns a vector of the size given which is filled with zeros if no random number generator is specified, if the optinal argument rng is supplied, it uses the generator to return a vector of random values of the given size.

POMDPDiscrete.policy_evaluation — Method

policy_evaluation(mdp, π)

Compute the value function by evaluating the policy in each state using the Bellman Expectation Equation. Sutton and Barto 2nd Edition p75 math V(s) = \sum_a\pi(a|s)\sum_{s^\prime, s} p(s^\prime, r |s, a) [r + \gamma v(s^\prime)]

POMDPDiscrete.policy_grid — Method

policy_grid(mdp, policy)
policy_grid(mdp, xmax, ymax)

Return a representation of the policy using unicode arrows stored in a Matrix{String}, where each state is represented by an element of the matrix. This only shows one action per state, samples distribution to select action.

POMDPDiscrete.policy_transition_matrix — Method

policy_transition_matrix(mdp, π)

Return a transition matrix given a probabilistic model Ps′, a, s) and a policy stochastic π[s, a].

POMDPDiscrete.random_stochastic_policy — Method

random_stochastic_policy(mdp)

Return a stationary StochasticPolicy` where the action distribution for each state is random.

POMDPDiscrete.uniform_stochastic_policy — Method

uniform_stochastic_policy(mdp)

Return a stationary StochasticPolicy where all actions are equally likely for each state.

POMDPDiscrete.value_iteration — Method

value_iteration(mdp)

Find the optimal value function for an agent acting greedily.

POMDPs.reward — Function

reward(mdp::GridWorld, s::State, [a::Any])

Returns a reward of -1 for any action a taken in state s for a GridWorld mdp. If the stsate s is in the set of absorbing states, then the reward is zero. Note the reward is function of the current state only.

POMDPs.transition — Method

transition(mdp::GridWorld, s::State, a::Symbol)

Returns the probability distribution of successor states for a GridWorld mdp, given the current state s and an action a selected in that state.